+++ CHANGELOG.md
@@ -1,47 +1,75 @@
 # Change Log
 
+## [1.0.0-RC10](https://github.com/the-AjK/btb/tree/1.0.0-RC10) (2018-04-19)
+
+[Full Changelog](https://github.com/the-AjK/btb/compare/1.0.0-RC9...1.0.0-RC10)
+
+### Changes
+
+- Added auto drink feature
+- Missing order user list
+
+### Bugfixes
+
+- Daily menu notification only for users who placed an order
+- Empty mentions handler
+
 ## [1.0.0-RC9](https://github.com/the-AjK/btb/tree/1.0.0-RC9) (2018-04-18)
+
 [Full Changelog](https://github.com/the-AjK/btb/compare/1.0.0-RC8...1.0.0-RC9)
 
 ### Changes
+
 - Added lunch rating
 
 ### Bugfixes
+
 - Various bugfixes
 
 ## [1.0.0-RC8](https://github.com/the-AjK/btb/tree/1.0.0-RC8) (2018-04-16)
+
 [Full Changelog](https://github.com/the-AjK/btb/compare/1.0.0-RC7...1.0.0-RC8)
 
 ### Changes
+
 - Added menu validation
 - Added menu create/update warnings
 
 ### Bugfixes
+
 - Menu create/update error
 - Various bugfixes
 
 ## [1.0.0-RC7](https://github.com/the-AjK/btb/tree/1.0.0-RC7) (2018-04-14)
+
 [Full Changelog](https://github.com/the-AjK/btb/compare/1.0.0-RC6...1.0.0-RC7)
 
 ### Changes
+
 - Orders server side pagination
 
 ### Bugfixes
+
 - Various bugfixes
 
 ## [1.0.0-RC6](https://github.com/the-AjK/btb/tree/1.0.0-RC6) (2018-04-13)
+
 [Full Changelog](https://github.com/the-AjK/btb/compare/1.0.0-RC5...1.0.0-RC6)
 
 ## [1.0.0-RC5](https://github.com/the-AjK/btb/tree/1.0.0-RC5) (2018-04-10)
+
 [Full Changelog](https://github.com/the-AjK/btb/compare/1.0.0-RC4...1.0.0-RC5)
 
 ## [1.0.0-RC4](https://github.com/the-AjK/btb/tree/1.0.0-RC4) (2018-04-09)
+
 [Full Changelog](https://github.com/the-AjK/btb/compare/1.0.0-RC3...1.0.0-RC4)
 
 ## [1.0.0-RC3](https://github.com/the-AjK/btb/tree/1.0.0-RC3) (2018-04-08)
+
 [Full Changelog](https://github.com/the-AjK/btb/compare/1.0.0-RC1...1.0.0-RC3)
 
 ## [1.0.0-RC1](https://github.com/the-AjK/btb/tree/1.0.0-RC1) (2018-04-05)
+
 [Full Changelog](https://github.com/the-AjK/btb/compare/1.0.0-RC0...1.0.0-RC1)
 
 ## [1.0.0-RC0](https://github.com/the-AjK/btb/tree/1.0.0-RC0) (2018-03-30)